requirements.txt
django==3.1.1
djangorestframework==3.11.1
django-extensions==3.0.9